Obsoleted _TASK_ROLLOVER_FIX
Based upon http://arduino.stackexchange.com/a/12588/10648, the extra
code for _TASK_ROLLOVER_FIX should not be needed if the math is done
slightly different. I updated the code and did some quick tests and it
appears correct. Example #6 returns the same values for IDLE, but a few
less ms when not compiled with IDLE. Using the setMillis() function in
the SO posting, I did some quick tests with Example #2. using
setMillis(-3000) sets the millis() value to 3 seconds before rollover.
Task #1 gets an extra catch-up hit, but not Task #2. But, it works the
same as the original code though when using setMillis(-3000).