delete manually disabled self-destruct tasks
adjusts an example to disable a single task before it runs out of
iterations. the task disables itself. without the fix in
TaskScheduler.h, one can observe a memory leak, as the task which
disabled itself is not deleted.

the TaskScheduler is adjusted to check the self-destruct request flag
of disabled tasks and to delete such tasks.